Fix certificate request for subdomains
When a certificate request was done for a subdomain, the code would try
to access the mijn.host API directly on that subdomain which would
result in an error code.

This commit adds code that queries the API multiple times to find the
correct base domain for the given subdomain.